% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforge.R
\name{rforge-mirror}
\alias{rforge-mirror}
\alias{rforge_mirror}
\alias{rforge_find_projects}
\alias{project name of the project (usually package)}
\alias{rforge_get_revision}
\alias{rforge_list_repos}
\alias{rforge_cleanup_repos}
\title{R-forge mirror tools}
\usage{
rforge_mirror(this_week = TRUE)

rforge_find_projects(this_week = TRUE)

rforge_get_revision(project)

rforge_list_repos()

rforge_cleanup_repos()
}
\arguments{
\item{this_week}{by default we only mirror repositories with activity in
the past week. Set this to FALSE to mirror all repos we can find.}
}
\description{
Mirrors SVN repositories to \url{https://github.com/r-forge}
}
